Change method of creation of ImmutableLists in ParserTest.cs
Previously, ImmutableLists were being created using the method
`new List<TYPE> { OBJECTS, IN, LIST }.ToImmutableList()`. This is
clearly a rather verbose method of creating an ImmutableList<T>. This
commit changes the method of creation of ImmutableLists to the more
concise, clearer `ImmutableList.Create(OBJECTS, IN, LIST). Note that
this commit only changes the method of creation in the file
ParserTest.cs; all other classes will be unaffected.