Change internal representation of algebraic data types
Algebraic data types are enumerations which can store data, along with
state; for instance, the algebraic data type variously known as `Maybe'
or `Option' can store a value in its `Just', or `Some', state, and
nothing in its `Nothing', or `None', state. Previously, MParse
represented algebraic data types using a class with:
- several private nested classes to represent each case,
- several private fields for storing the current value of the algebraic
  data type,
- several private properties that acted as `frontends' to the fields,
  and
- several factory methods, one for each case.
This was clearly, to say the least, an inelegant and clumsy solution.
A more elegant solution, found at http://bugsquash.blogspot.com.au/2012/
01/encoding-algebraic-data-types-in-c.html, represents algebraic data
types in C# using a class with:
- a nested, sealed subclass of the containing class for each case, and
- a private constructor so that the class is `sealed' (no more
  subclasses can be created).